Update imports after moving runner.ts
