+++ src/util/runner.ts
@@ -1,6 +1,6 @@
-import {streamOutput} from '../../util/child-process'
+import {streamOutput} from '../util/child-process'
 import {SpawnOptions} from 'child_process'
-import {Terraform} from '../../tools/terraform'
+import {Terraform} from '../tools/terraform'
 
 export async function runner(argv: string[], spawnOptions?: SpawnOptions) {
   const op: SpawnOptions = spawnOptions || {stdio: ['ignore',  'pipe', 'pipe']}
